Generación de pedidos automática.

+++ src/main/java/com/gifa_api/dto/proveedoresYPedidos/AsociacionProveedorDeITemDTO.java
@@ -16,5 +16,5 @@ public class AsociacionProveedorDeITemDTO {
 
     Integer idItem;
     Integer idProveedor;
-    BigDecimal precio;
+    Double precio;
 }
